% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello}
\alias{say_hello}
\title{Function to say hello to a given person}
\usage{
say_hello(someone = "You")
}
\arguments{
\item{someone}{Character Name of the person to say hello to}
}
\value{
Side effect: Display message in the command to say hello to someone
}
\description{
Function to say hello to a given person
}
\examples{
say_hello() # Hello You
say_hello(someone = "World") # Hello World
}
